+++ Realization/backend/acc/src/main/java/eu/bcvsolutions/idm/acc/service/api/SysSystemFormValueService.java
@@ -3,16 +3,17 @@ package eu.bcvsolutions.idm.acc.service.api;
 import java.util.UUID;
 
 import eu.bcvsolutions.idm.acc.entity.SysSystem;
-import eu.bcvsolutions.idm.acc.entity.SysSystemFormValue;
 import eu.bcvsolutions.idm.core.api.service.CloneableService;
-import eu.bcvsolutions.idm.core.eav.service.api.FormValueService;
+import eu.bcvsolutions.idm.core.eav.api.dto.IdmFormValueDto;
+import eu.bcvsolutions.idm.core.eav.api.service.FormValueService;
 
 /**
  * Service for control form value on system
+ * 
  * @author svandav
  *
  */
-public interface SysSystemFormValueService extends FormValueService<SysSystem, SysSystemFormValue>, CloneableService<SysSystemFormValue> {
+public interface SysSystemFormValueService extends FormValueService<SysSystem>, CloneableService<IdmFormValueDto> {
 
 	/**
 	 * Duplicate form value. Create and persist new form value (included confidential value save)
@@ -20,6 +21,6 @@ public interface SysSystemFormValueService extends FormValueService<SysSystem, S
 	 * @param owner for this value
 	 * @return
 	 */
-	SysSystemFormValue duplicate(UUID id, SysSystem owner);
+	IdmFormValueDto duplicate(UUID id, SysSystem owner);
 
 }