+++ Realization/backend/acc/src/main/java/eu/bcvsolutions/idm/acc/service/impl/RoleSynchronizationExecutor.java
@@ -52,13 +52,12 @@ import eu.bcvsolutions.idm.core.api.dto.filter.CorrelationFilter;
 import eu.bcvsolutions.idm.core.api.dto.filter.IdmRoleFilter;
 import eu.bcvsolutions.idm.core.api.entity.AbstractEntity;
 import eu.bcvsolutions.idm.core.api.event.EntityEvent;
-import eu.bcvsolutions.idm.core.api.repository.BaseEntityRepository;
 import eu.bcvsolutions.idm.core.api.service.ConfidentialStorage;
 import eu.bcvsolutions.idm.core.api.service.EntityEventManager;
 import eu.bcvsolutions.idm.core.api.service.GroovyScriptService;
 import eu.bcvsolutions.idm.core.api.service.ReadWriteDtoService;
 import eu.bcvsolutions.idm.core.eav.api.entity.FormableEntity;
-import eu.bcvsolutions.idm.core.eav.service.api.FormService;
+import eu.bcvsolutions.idm.core.eav.api.service.FormService;
 import eu.bcvsolutions.idm.core.model.entity.IdmRole;
 import eu.bcvsolutions.idm.core.model.event.RoleEvent;
 import eu.bcvsolutions.idm.core.model.event.RoleEvent.RoleEventType;
@@ -380,11 +379,6 @@ public class RoleSynchronizationExecutor extends AbstractSynchronizationExecutor
 	@Override
 	protected Class<? extends FormableEntity> getEntityClass() {
 		return IdmRole.class;
-	}
-
-	@Override
-	protected BaseEntityRepository getRepository() {
-		return this.roleRepository;
 	}
 
 	@Override