+++ Realization/backend/core/core-impl/src/main/java/eu/bcvsolutions/idm/core/model/repository/eav/IdmRoleFormValueRepository.java
@@ -1,7 +1,5 @@
 package eu.bcvsolutions.idm.core.model.repository.eav;
 
-import org.springframework.data.rest.core.annotation.RepositoryRestResource;
-
 import eu.bcvsolutions.idm.core.eav.repository.AbstractFormValueRepository;
 import eu.bcvsolutions.idm.core.model.entity.IdmRole;
 import eu.bcvsolutions.idm.core.model.entity.eav.IdmRoleFormValue;
@@ -12,11 +10,6 @@ import eu.bcvsolutions.idm.core.model.entity.eav.IdmRoleFormValue;
  * @author Radek Tomiška
  *
  */
-@RepositoryRestResource(
-		itemResourceRel = "formValue",
-		collectionResourceRel = "formValues",
-		exported = false
-		)
 public interface IdmRoleFormValueRepository extends AbstractFormValueRepository<IdmRole, IdmRoleFormValue> {
 	
 }